+++ backend/server.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/joho/godotenv"
-	"github.com/kennedynguyen1/swipe-rank/backend/webhooks"
+	webhooks "github.com/kennedynguyen1/swipe-rank/backend/webhook"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
